Make configureable which Brightness IDs are tried
"It would be better to have it stored in the settings instead as a list,
I haven't had time to figure that out. A dropdown list perhaps with
multiselect values, 10 is selected and has 6B unselected. If you have
time to do those then it would be easy to merge" -- Binay Devkota

from https://github.com/daitj/gnome-display-brightness-ddcutil/pull/152