DRAFT: writeCollection: Wait until previous invocation finished
This commit changes how writeCollection works. Previously, it seemed
to be a forced delay of some number of milliseconds that would need to
happen before a call to ddcutil to change brightness could occur.

Update to use a simpler logic: No more delays. If a ddcutil set
brightness command is running, queue the next command that's generated
and run it immediately after the first command completes. If a command
is queued and another one comes in, the newest one replaces the queued
one, becoming the only queued one.

This patch is a draft. If this concept is acceptable, a futher update
would be needed to remove the concept of writeCollectorWaitMs.
Alternatively, we could keep the concept and mandate a minimum
interval between two invocationsn of ddcutil, either measured from
start time to start time, or end time to start time.